:truck: Rename Alexa resources
