--- platforms/platform-alexa/src/AlexaSkill.ts
@@ -1,12 +1,0 @@
-import { Jovo } from '@jovotech/framework';
-import { AlexaResponse } from '@jovotech/output-alexa';
-import { AlexaRequest } from './AlexaRequest';
-
-export class AlexaSkill extends Jovo<AlexaRequest, AlexaResponse> {
-  getSkillId(): string | undefined {
-    return (
-      this.$request.session?.application?.applicationId ||
-      this.$request.context?.System?.application?.applicationId
-    );
-  }
-}
