+++ platforms/platform-alexa/src/index.ts
@@ -1,31 +1,31 @@
 import { registerPlatformSpecificJovoReference, isNode } from '@jovotech/framework';
-import { Alexa, AlexaConfig } from './Alexa';
-import { AlexaSkill } from './AlexaSkill';
+import { AlexaPlatform, AlexaConfig } from './AlexaPlatform';
+import { Alexa } from './Alexa';
 import type { AlexaCli as AlexaCliType } from './cli';
 
 declare module '@jovotech/framework/dist/types/Extensible' {
   interface ExtensiblePluginConfig {
-    Alexa?: AlexaConfig;
+    AlexaPlatform?: AlexaConfig;
   }
 
   interface ExtensiblePlugins {
-    Alexa?: Alexa;
+    AlexaPlatform?: AlexaPlatform;
   }
 }
 
 declare module '@jovotech/framework/dist/types/Jovo' {
   interface Jovo {
-    $alexaSkill?: AlexaSkill;
+    $alexa?: Alexa;
   }
 }
-registerPlatformSpecificJovoReference('$alexaSkill', AlexaSkill);
+registerPlatformSpecificJovoReference('$alexa', Alexa);
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 export const AlexaCli: typeof AlexaCliType = isNode() ? require('./cli').AlexaCli : null;
 export * from './Alexa';
+export * from './AlexaPlatform';
 export * from './AlexaRequest';
 export * from './AlexaUser';
-export * from './AlexaSkill';
 export type { AlexaResponse } from '@jovotech/output-alexa';
 export * from './constants';
 export * from './interfaces';
